feat: [TagsObObjectDefinitionTrait] Creates php-trait for all tag operations on php class.
