+++ src/DiContainer/DiDefinition/DiDefinitionAutowire.php
@@ -4,14 +4,11 @@ declare(strict_types=1);
 
 namespace Kaspi\DiContainer\DiDefinition;
 
-use InvalidArgumentException;
 use Kaspi\DiContainer\AttributeReader;
 use Kaspi\DiContainer\Attributes\Setup;
-use Kaspi\DiContainer\Attributes\Tag;
 use Kaspi\DiContainer\DiDefinition\Arguments\ArgumentBuilder;
 use Kaspi\DiContainer\DiDefinition\Arguments\ArgumentResolver;
 use Kaspi\DiContainer\Enum\SetupConfigureMethod;
-use Kaspi\DiContainer\Exception\AutowireException;
 use Kaspi\DiContainer\Exception\DiDefinitionException;
 use Kaspi\DiContainer\Helper;
 use Kaspi\DiContainer\Interfaces\DiContainerInterface;
@@ -21,43 +18,33 @@ use Kaspi\DiContainer\Interfaces\DiDefinition\DiDefinitionAutowireInterface;
 use Kaspi\DiContainer\Interfaces\DiDefinition\DiDefinitionIdentifierInterface;
 use Kaspi\DiContainer\Interfaces\DiDefinition\DiDefinitionSetupAutowireInterface;
 use Kaspi\DiContainer\Interfaces\DiDefinition\DiDefinitionTagArgumentInterface;
-use Kaspi\DiContainer\Interfaces\DiDefinition\DiTaggedDefinitionInterface;
+use Kaspi\DiContainer\Interfaces\DiDefinition\DiTaggedObjectDefinitionInterface;
 use Kaspi\DiContainer\Interfaces\Exceptions\DiDefinitionExceptionInterface;
 use Kaspi\DiContainer\Traits\BindArgumentsTrait;
-use Kaspi\DiContainer\Traits\TagsTrait;
+use Kaspi\DiContainer\Traits\TagsOnObjectDefinitionTrait;
 use ReflectionClass;
 use ReflectionException;
 
 use function call_user_func_array;
 use function get_class;
 use function get_debug_type;
-use function is_callable;
-use function is_int;
-use function is_null;
 use function is_object;
 use function is_string;
 use function sprintf;
-use function trim;
-use function var_export;
 
 /**
  * @phpstan-import-type DiDefinitionType from DiDefinitionArgumentsInterface
- * @phpstan-import-type Tags from DiTaggedDefinitionInterface
- * @phpstan-import-type TagOptions from DiDefinitionTagArgumentInterface
  *
  * @phpstan-type SetupConfigureArgumentsType array<non-empty-string|non-negative-int, DiDefinitionType|mixed>
  * @phpstan-type SetupConfigureItem array{0: SetupConfigureMethod, 1: SetupConfigureArgumentsType}
  */
-final class DiDefinitionAutowire implements DiDefinitionAutowireInterface, DiDefinitionSetupAutowireInterface, DiDefinitionIdentifierInterface, DiDefinitionTagArgumentInterface, DiTaggedDefinitionInterface
+final class DiDefinitionAutowire implements DiDefinitionAutowireInterface, DiDefinitionSetupAutowireInterface, DiDefinitionIdentifierInterface, DiDefinitionTagArgumentInterface, DiTaggedObjectDefinitionInterface
 {
     use BindArgumentsTrait {
         bindArguments as private bindArgumentsInternal;
     }
-    use TagsTrait {
-        getTags as public getBoundTags;
-        hasTag as private hasTagInternal;
-        geTagPriority as private geTagPriorityInternal;
-    }
+
+    use TagsOnObjectDefinitionTrait;
 
     private ReflectionClass $reflectionClass;
 
@@ -67,13 +54,6 @@ final class DiDefinitionAutowire implements DiDefinitionAutowireInterface, DiDef
      * @var list<array{0: SetupConfigureMethod, 1: ArgumentBuilderInterface}>
      */
     private array $setupArgBuilders;
-
-    /**
-     * Tags from php attributes on class.
-     *
-     * @var array<non-empty-string, TagOptions>
-     */
-    private array $tagsByAttribute;
 
     /**
      * Methods for setup service by PHP definition via setters (mutable or immutable).
@@ -88,8 +68,6 @@ final class DiDefinitionAutowire implements DiDefinitionAutowireInterface, DiDef
      * @var array<non-empty-string, list<SetupConfigureItem>>
      */
     private array $setupByAttributes;
-
-    private DiContainerInterface $container;
 
     /**
      * @param class-string|ReflectionClass $definition
@@ -131,13 +109,6 @@ final class DiDefinitionAutowire implements DiDefinitionAutowireInterface, DiDef
     public function isSingleton(): ?bool
     {
         return $this->isSingleton;
-    }
-
-    public function setContainer(DiContainerInterface $container): static
-    {
-        $this->container = $container;
-
-        return $this;
     }
 
     public function exposeArgumentBuilder(DiContainerInterface $container): ?ArgumentBuilderInterface
@@ -227,7 +198,7 @@ final class DiDefinitionAutowire implements DiDefinitionAutowireInterface, DiDef
     }
 
     /**
-     * @return class-string|non-empty-string
+     * @return class-string
      */
     public function getIdentifier(): string
     {
@@ -236,145 +207,14 @@ final class DiDefinitionAutowire implements DiDefinitionAutowireInterface, DiDef
             : $this->reflectionClass->getName();
     }
 
-    /**
-     * @throws DiDefinitionExceptionInterface
-     */
-    public function getTags(): array
+    public function isImplementInterface(string $interface): bool
     {
-        if (!$this->getContainer()->getConfig()->isUseAttribute()) {
-            return $this->getBoundTags();
-        }
-
-        try {
-            // 🚩 PHP attributes have higher priority than PHP definitions (see documentation.)
-            return $this->getTagsByAttribute() + $this->getBoundTags();
-        } catch (DiDefinitionExceptionInterface $e) {
-            throw new DiDefinitionException(
-                message: sprintf('Cannot get tags on class "%s".', $this->getIdentifier()),
-                previous: $e,
-            );
-        }
+        return $this->getDefinition()->implementsInterface($interface);
     }
 
-    /**
-     * @throws DiDefinitionExceptionInterface
-     */
-    public function hasTag(string $name): bool
+    public function getDefinitionIdentifier(): string
     {
-        try {
-            return isset($this->getTags()[$name]);
-        } catch (DiDefinitionExceptionInterface $e) {
-            throw new DiDefinitionException(
-                message: sprintf('Cannot check exist tag "%s" on class "%s".', $name, $this->getIdentifier()),
-                previous: $e,
-            );
-        }
-    }
-
-    /**
-     * @throws DiDefinitionExceptionInterface
-     */
-    public function getTag(string $name): ?array
-    {
-        try {
-            return $this->getTags()[$name] ?? null;
-        } catch (DiDefinitionExceptionInterface $e) {
-            throw new DiDefinitionException(
-                message: sprintf('Cannot get tag "%s" on class "%s".', $name, $this->getIdentifier()),
-                previous: $e,
-            );
-        }
-    }
-
-    /**
-     * @param non-empty-string $name
-     * @param TagOptions       $operationOptions
-     *
-     * @throws DiDefinitionExceptionInterface
-     */
-    public function geTagPriority(string $name, array $operationOptions = []): int|string|null
-    {
-        if (null !== ($priority = $this->geTagPriorityInternal($name, $operationOptions))) {
-            return $priority;
-        }
-
-        $tagOptions = $operationOptions + ($this->getTag($name) ?? []);
-
-        if (isset($tagOptions['priority.method'])) {
-            $method = $tagOptions['priority.method'];
-
-            if (!is_string($method) || '' === trim($method)) {
-                $wherePriorityMethod = isset($this->getBoundTags()[$name]['priority.method'])
-                    ? 'value with key "priority.method" in the $options parameter in '.DiDefinitionTagArgumentInterface::class.'::bindTag()'
-                    : 'the $priorityMethod parameter or the value with key "priority.method" in the $options parameter in the php attribute #[Tag]';
-
-                throw new DiDefinitionException(
-                    sprintf('Cannot get tag priority for tag "%s" via method in class %s. The name of the priority method is specified by %s. Priority method must be present none-empty string. Got: %s', $name, $this->getIdentifier(), $wherePriorityMethod, var_export($method, true))
-                );
-            }
-
-            try {
-                return $this->getTagPriorityFromMethod($method, $name, $tagOptions);
-            } catch (AutowireException|InvalidArgumentException $e) {
-                throw new DiDefinitionException(
-                    message: sprintf('Cannot get tag priority for tag "%s" via method %s::%s(). Caused by: %s', $name, $this->getIdentifier(), $method, $e->getMessage()),
-                    previous: $e
-                );
-            }
-        }
-
-        // Option (meta-key) only from $operationOptions. It uses through DiDefinitionTaggedAs.
-        $priorityDefaultMethod = ($operationOptions['priority.default_method'] ?? null);
-
-        if (!is_string($priorityDefaultMethod)) {
-            return null;
-        }
-
-        try {
-            return $this->getTagPriorityFromMethod($priorityDefaultMethod, $name, $tagOptions);
-        } catch (InvalidArgumentException) {
-            return null;
-        } catch (AutowireException $e) {
-            throw new DiDefinitionException(
-                message: sprintf('Cannot get tag priority for tag "%s" via default priority method %s::%s(). Caused by: %s', $name, $this->getIdentifier(), $priorityDefaultMethod, $e->getMessage()),
-                previous: $e
-            );
-        }
-    }
-
-    /**
-     * @return array<non-empty-string, TagOptions>
-     *
-     * @throws DiDefinitionExceptionInterface
-     */
-    public function getTagsByAttribute(): array
-    {
-        if (isset($this->tagsByAttribute)) {
-            return $this->tagsByAttribute;
-        }
-
-        $this->tagsByAttribute = [];
-
-        try {
-            $tagAttributes = AttributeReader::getTagAttribute($this->getDefinition());
-        } catch (DiDefinitionExceptionInterface $e) {
-            throw new DiDefinitionException(
-                message: sprintf('Cannot read php attribute #[%s] on class "%s".', Tag::class, $this->getIdentifier()),
-                previous: $e,
-            );
-        }
-
-        foreach ($tagAttributes as $tagAttribute) {
-            $priorityMethod = null !== $tagAttribute->priorityMethod
-                ? ['priority.method' => $tagAttribute->priorityMethod]
-                : [];
-            $this->tagsByAttribute[$tagAttribute->name] = self::transformTagOptions(
-                $priorityMethod + $tagAttribute->options,
-                $tagAttribute->priority
-            );
-        }
-
-        return $this->tagsByAttribute;
+        return $this->getIdentifier();
     }
 
     /**
@@ -401,17 +241,6 @@ final class DiDefinitionAutowire implements DiDefinitionAutowireInterface, DiDef
         return $this->setupByAttributes + $this->setup;
     }
 
-    private function getContainer(): DiContainerInterface
-    {
-        if (!isset($this->container)) {
-            throw new DiDefinitionException(
-                sprintf('Need set container implementation. Use method %s::setContainer(). Definition identifier "%s".', __CLASS__, $this->getIdentifier())
-            );
-        }
-
-        return $this->container;
-    }
-
     /**
      * @throws DiDefinitionExceptionInterface
      */
@@ -420,29 +249,5 @@ final class DiDefinitionAutowire implements DiDefinitionAutowireInterface, DiDef
         if (!$this->getDefinition()->isInstantiable()) {
             throw new DiDefinitionException(sprintf('The "%s" class is not instantiable.', $this->getDefinition()->getName()));
         }
-    }
-
-    /**
-     * @param TagOptions $tagOptions
-     *
-     * @throws AutowireException|InvalidArgumentException
-     */
-    private function getTagPriorityFromMethod(string $method, string $tag, array $tagOptions): int|string|null
-    {
-        $callable = [$this->getIdentifier(), $method];
-
-        if (!is_callable($callable)) {
-            throw new InvalidArgumentException('Method must be declared with public and static modifiers.');
-        }
-
-        $priority = $callable($tag, $tagOptions);
-
-        if (is_int($priority) || is_string($priority) || is_null($priority)) {
-            return $priority;
-        }
-
-        throw new AutowireException(
-            sprintf('Method must return type "int|string|null" but return type "%s".', get_debug_type($priority))
-        );
     }
 }
